pyproject.toml
# SPDX-FileCopyrightText: 2024 Lukas Schrangl <lukas.schrangl@boku.ac.at>
#
# SPDX-License-Identifier: BSD-3-Clause

[project]
name = "smfret-bondtime"
description = "Infer receptor–ligand bond lifetimes from smFRET data"
authors = [
    { name = "Lukas Schrangl", email = "lukas.schrangl@boku.ac.at" },
]
maintainers = [
    { name = "Lukas Schrangl", email = "lukas.schrangl@boku.ac.at" },
]
requires-python = ">=3.10"
dependencies = [
    "PyQt5",
    "lifelines",
    "matplotlib",
    "numpy",
    "pandas",
    "scipy",
    "sdt-python>=18.3",
    "trackpy",
]
dynamic = ["version"]

[project.gui-scripts]
smfret-bondtime = "smfret_bondtime.gui:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.urls]
repository = "https://github.com/schuetzgroup/smfret-bondtime"

[tool.hatch.version]
path = "smfret_bondtime/_version.py"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
